+++ python/interpret-core/interpret/develop/__init__.py
@@ -33,7 +33,8 @@ def debug_info():
     Returns:
         A dictionary that contains debug info across the interpret package.
     """
-    from . import __version__, status_show_server
+    from interpret.version import __version__
+    from interpret.visual import status_show_server
 
     debug_dict = {}
     debug_dict["interpret.__version__"] = __version__
@@ -141,7 +142,7 @@ def debug_mode(log_filename="log.txt", log_level="INFO", native_debug=True):
     """
     import json
     import logging
-    from .glassbox.ebm.internal import Native
+    from interpret.glassbox.ebm.internal import Native
 
     # Exit fast on second call.
     if this.is_debug_mode: