+++ python/interpret-core/interpret/glassbox/ebm/test/test_internal.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2019 Microsoft Corporation
 # Distributed under the MIT software license
 
-from ..internal import Native, NativeEBMBooster
+from interpret.glassbox.ebm.internal import Native, NativeEBMBooster
 
 import numpy as np
 import ctypes as ct