+++ python/interpret-core/interpret/perf/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2019 Microsoft Corporation
 # Distributed under the MIT software license
 
-from .curve import ROC, PR  # noqa: F401
-from .regression import RegressionPerf  # noqa: F401
+from interpret.perf.curve import ROC, PR  # noqa: F401
+from interpret.perf.regression import RegressionPerf  # noqa: F401