+++ python/interpret-core/interpret/visual/__init__.py
@@ -1,2 +1,15 @@
 # Copyright (c) 2019 Microsoft Corporation
 # Distributed under the MIT software license
+
+from interpret.visual.interactive import (  # noqa: F401
+    show,
+    show_link,
+    set_show_addr,
+    get_show_addr,
+)
+from interpret.visual.interactive import preserve  # noqa: F401
+from interpret.visual.interactive import shutdown_show_server  # noqa: F401
+from interpret.visual.interactive import init_show_server  # noqa: F401
+from interpret.visual.interactive import status_show_server  # noqa: F401
+from interpret.visual.interactive import set_visualize_provider  # noqa: F401
+from interpret.visual.interactive import get_visualize_provider  # noqa: F401