+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and the versioning is mostly derived from [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.3.0] - TBD
+### Breaking Changes
+- Native namespacing is used for top-level `interpret` namespace.
+- Namespace `interpret` has all top-level methods moved to their own modules.
+  - ```
+    interpret.show -> interpret.visual.show
+    interpret.show_link -> interpret.visual.show_link
+    interpret.set_show_addr -> interpret.visual.set_show_addr
+    interpret.get_show_addr -> interpret.visual.get_show_addr
+    interpret.preserve -> interpret.visual.preserve
+    interpret.shutdown_show_server -> interpret.visual.shutdown_show_server
+    interpret.init_show_server -> interpret.visual.init_show_server
+    interpret.status_show_server -> interpret.visual.status_show_server
+    interpret.set_visualize_provider -> interpret.visual.set_visualize_provider
+    interpret.get_visualize_provider -> interpret.visual.get_visualize_provider
+
+    interpret.__version__ -> interpret.version.__version__
+    ```
+- NullHandler logger for `interpret` package removed.
+
 ## [v0.2.5] - 2021-06-21
 ### Added
 - Sample weight support added for EBM.